refactoring done, unsure about export const aboutme part
