+++ native/screens/UserProfile/UserProfile.jsx
@@ -1,13 +1,12 @@
 import { useState, useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import Ionicons from 'react-native-vector-icons/Ionicons';
-import { ShowPortfolioProjects } from '../components/PortfolioCard/ShowPortfolioProjects.jsx';
-import { uiActions } from '../services/redux/slices/uiSlice';
-import { getOneUser } from '../services/api/users';
-import { Button, Linking, Modal, ScrollView, StyleSheet, Text, View } from 'react-native';
-import { handleTextChange } from '../services/utils/handlers.js';
-import { updateUserAndProject } from '../services/api/projects.js';
-import { fetchAllProjects } from '../services/redux/actions/projectActions.js';
+import { uiActions } from '../../services/redux/slices/uiSlice';
+import { getOneUser } from '../../services/api/users';
+import { Linking, ScrollView, Text, View } from 'react-native';
+import { handleTextChange } from '../../services/utils/handlers.js';
+import { styles } from './styles';
+import { OwnerOptions, DecisionModal } from './helpers.jsx';
 
 export const UserProfile = ({ route, navigation }) => {
   const [reviewStatus, setReviewStatus] = useState({
@@ -141,154 +140,3 @@ export const UserProfile = ({ route, navigation }) => {
     </ScrollView>
   );
 };
-
-const OwnerOptions = ({ setReviewStatus }) => {
-  return (
-    <View>
-      <Text>{'REQUEST TO JOIN MESSAGE GOES HEREN LOREM IPSUM LATINN MORE LATIN IPSUM LOREM'}</Text>
-      <Button
-        title="Approve Request"
-        onPress={() => {
-          handleTextChange(true, 'renderDecisionModal', setReviewStatus);
-          handleTextChange(true, 'ownerDecision', setReviewStatus);
-        }}
-      />
-      <Button
-        title="Reject Request"
-        onPress={() => {
-          handleTextChange(true, 'renderDecisionModal', setReviewStatus);
-          handleTextChange(false, 'ownerDecision', setReviewStatus);
-        }}
-      />
-    </View>
-  );
-};
-
-const DecisionModal = ({ applicant, navigation, reviewStatus, setReviewStatus }) => {
-  const { _id: userID, first_name, interested_projects } = applicant;
-  const { ownerDecision, renderDecisionModal } = reviewStatus;
-  const {
-    _id: projectID,
-    team_members: projectMembers,
-    interested_applicants,
-  } = reviewStatus.ownerProject;
-
-  let requestBody = {
-    project: {
-      projectID,
-      projectUpdate: {
-        // interested applicants has full applicant object
-        interested_applicants: interested_applicants.filter((applicant) => applicant._id != userID),
-      },
-    },
-    user: {
-      userID,
-      userUpdate: {
-        // interested proejcts ONLY has project ID
-        interested_projects: interested_projects.filter(
-          (interestedProjectID) => interestedProjectID != projectID
-        ),
-      },
-    },
-  };
-
-  useEffect(() => {
-    if (ownerDecision) {
-      setTimeout(() => {
-        addToTeam();
-        endReviewSequence();
-      }, 5000);
-    }
-  }, [ownerDecision]);
-
-  const endReviewSequence = () => {
-    handleTextChange(false, 'renderDecisionModal', setReviewStatus);
-    handleTextChange(false, 'ownerViewingApplicant', setReviewStatus);
-    navigation.navigate('Applicants', {
-      projectID,
-    });
-  };
-
-  const rejectApplicant = async () => {
-    await updateUserAndProject(requestBody);
-    fetchAllProjects();
-    endReviewSequence();
-  };
-
-  const addToTeam = async () => {
-    requestBody = {
-      ...requestBody,
-      project: {
-        ...requestBody.project,
-        team_members: [...projectMembers, userID],
-      },
-    };
-    await updateUserAndProject(requestBody);
-    fetchAllProjects();
-  };
-
-  return (
-    <Modal visible={renderDecisionModal}>
-      <View>
-        {ownerDecision ? (
-          <Text>{`${first_name} has been added to your project!`}</Text>
-        ) : (
-          <>
-            <Text>{`Reject ${first_name}`}</Text>
-            <Button title="Yes, reject." onPress={rejectApplicant} />
-            <Button
-              title="Cancel"
-              onPress={() => handleTextChange(false, 'renderDecisionModal', setReviewStatus)}
-            />
-          </>
-        )}
-      </View>
-    </Modal>
-  );
-};
-
-const styles = StyleSheet.create({
-  editContainer: {
-    alignItems: 'flex-end',
-    marginRight: 5,
-  },
-  editButton: {
-    fontSize: 45,
-    marginRight: 10,
-  },
-  name: {
-    fontSize: 20,
-    fontWeight: 'bold',
-    marginTop: 10,
-  },
-  nameContainer: {
-    alignItems: 'center',
-    marginTop: 10,
-  },
-  placeholderImage: {
-    borderRadius: 100,
-    backgroundColor: '#c4c4c4',
-    height: 150,
-    width: 150,
-  },
-  portfolio: {
-    color: 'white',
-    fontSize: 35,
-    textAlign: 'center',
-    textDecorationLine: 'underline',
-  },
-  title: {
-    color: 'white',
-    fontWeight: 'bold',
-    marginLeft: 10,
-  },
-  words: {
-    color: 'white',
-    marginHorizontal: 10,
-  },
-  wordContainer: {
-    backgroundColor: '#313131',
-    marginHorizontal: 15,
-    marginTop: 15,
-  },
-});
